requirements.txt
asgiref==3.6.0
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cinemagoer==2022.12.27
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
drf-writable-nested==0.7.0
drf-yasg==1.21.4
exceptiongroup==1.0.4
greenlet==2.0.1
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==22.0
Pillow==9.4.0
pluggy==1.0.0
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
